client: fix warnings
client/Client.cc: In member function 'int Client::_read(Fh*, int64_t, uint64_t, ceph::bufferlist*)':
warning: client/Client.cc:5893:27: comparison between signed and unsigned integer expressions [-Wsign-compare]
client/Client.cc: In member function 'int Client::_write(Fh*, int64_t, uint64_t, const char*)':
warning: client/Client.cc:6235:30: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ src/client/Client.cc
@@ -5890,7 +5890,7 @@ int Client::_read(Fh *f, int64_t offset, uint64_t size, bufferlist *bl)
           bl->substr_of(in->inline_data, offset, len - offset);
           bl->append_zero(endoff - len);
         }
-      } else if (offset < endoff) {
+      } else if ((uint64_t)offset < endoff) {
         bl->append_zero(endoff - offset);
       }
